requirements.txt
fastapi
uvicorn[standard]
sqlmodel
alembic
psycopg2-binary
python-dotenv
pytest
httpx
pydantic[email]
